memusage: move MallocUsage into separate file
That way it becomes usable in other code like allocators/pool
