+++ src/malloc_usage.h
@@ -1,0 +1,30 @@
+// Copyright (c) 2023-present The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#ifndef BITCOIN_MALLOC_USAGE_H
+#define BITCOIN_MALLOC_USAGE_H
+
+#include <cassert>
+#include <cstddef>
+
+namespace memusage {
+
+/** Compute the total memory used by allocating alloc bytes. */
+static inline size_t MallocUsage(size_t alloc)
+{
+    // Measured on libc6 2.19 on Linux.
+    if (alloc == 0) {
+        return 0;
+    } else if (sizeof(void*) == 8) {
+        return ((alloc + 31) >> 4) << 4;
+    } else if (sizeof(void*) == 4) {
+        return ((alloc + 15) >> 3) << 3;
+    } else {
+        assert(0);
+    }
+}
+
+} // namespace memusage
+
+#endif // BITCOIN_MALLOC_USAGE_H
